requirements.txt
python-telegram-bot
timber==1.0.5
pika==0.12.0
PyYAML==3.13
requests==2.20.1
sentry-sdk==0.6.1
suds-jurko==0.6
cython==0.28.3